requirements.txt
Flask
flask-cors
flask-limiter
google-generativeai
fuzzywuzzy
requests
dns.resolver
tldextract
whois
beautifulsoup4
cryptography
pyaudio
gtts
pygame
numpy
transformers
torch
Pillow
scipy
opencv-python
sklearn
